+++ setup.sh
@@ -3,7 +3,10 @@
 ############
 ### Vars ###
 ############
-source_dir=$(dirname "$(readlink -f "$0")")
+source_dir=$(pushd $(dirname $0) >/dev/null && pwd && popd >/dev/null)
+
+is_linux="$(uname -s | grep Linux)"
+is_macos="$(uname -s | grep Darwin)"
 
 has_head="$DISPLAY"
 is_in_docker=$(awk -F/ '$2 == "docker"' /proc/self/cgroup)
@@ -12,8 +15,14 @@ export enable_services="systemctl --now enable"
 ################
 ### Packages ###
 ################
-$source_dir/distros/manjaro/setup.sh
-$source_dir/distros/ubuntu/setup.sh
+if [ $is_linux ]; then
+    $source_dir/distros/manjaro/setup.sh
+    $source_dir/distros/ubuntu/setup.sh
+fi
+
+if [ $is_macos ]; then
+    $source_dir/distros/macos/setup.sh
+fi
 
 mkdir -p ~/.terminfo/x
 wget 'https://github.com/kovidgoyal/kitty/blob/master/terminfo/x/xterm-kitty?raw=true' -qO ~/.terminfo/x/xterm-kitty
@@ -30,19 +39,23 @@ mkdir -p "$HOME/.local/share/Steam/steamapps/common/Counter-Strike Global Offens
 stow home -t $HOME -R -d $source_dir --adopt
 
 if [ ! "$is_in_docker" ]; then
-    sudo rm -rf /etc/ssh/sshd_config
+    if [ $is_linux ]; then
+        sudo rm -rf /etc/ssh/sshd_config
+    fi
     sudo stow etc -t /etc -R -d $source_dir --adopt
 fi
 
 if [ $has_head ]; then
-    cat $source_dir/gnome-settings.ini | dconf load /
-    sudo -u gdm dbus-launch gsettings set org.gnome.settings-daemon.plugins.power sleep-inactive-ac-type 'nothing'
+    if [ $is_linux ]; then
+        cat $source_dir/gnome-settings.ini | dconf load /
+        sudo -u gdm dbus-launch gsettings set org.gnome.settings-daemon.plugins.power sleep-inactive-ac-type 'nothing'
+    fi
 fi
 
 #############
 ### Shell ###
 #############
-sudo chsh -s $(which fish) $USER
+sudo chsh $USER -s $(which fish)
 
 wget -q https://raw.githubusercontent.com/oh-my-fish/oh-my-fish/master/bin/install
 fish install --noninteractive --yes
@@ -52,13 +65,15 @@ rm -f install
 ### Services ###
 ################
 if [ ! "$is_in_docker" ]; then
-    sudo gpasswd -a $USER docker
-    $enable_services docker.service
+    if [ $is_linux ]; then
+        sudo gpasswd -a $USER docker
+        $enable_services docker.service
 
-    $enable_services sshd.service
+        $enable_services sshd.service
 
-    sudo ufw limit ssh
-    yes | sudo ufw enable
+        sudo ufw limit ssh
+        yes | sudo ufw enable
+    fi
 
     if [ $has_head ]; then
         $enable_services --user randwall.service
